shared/python/requirements.txt
# Core dependencies for playground_sdk

# HTTP client for service calls
httpx>=0.25.0

# OpenTelemetry for tracing
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-cloud-trace>=1.6.0

# Utilities
python-dotenv>=1.0.0
